cable_guy: take types out of manager.py
